댓글 수정삭제 작업중

+++ src/main/java/com/meta/ponkids/domain/system/ntt/repository/NttRepository.java
@@ -30,7 +30,7 @@ public interface NttRepository extends JpaRepository< Ntt, Long> , NttRepository
            + "        , updt_dt = now() "
            + "    WHERE ntt_sn = :nttSn", nativeQuery = true )
        // nativeQuery true 없으면 error
-   int deleteAllByNttSn( @Param( "nttSn" ) int bbsSn );
+   int deleteAllByNttSn( @Param( "nttSn" ) Long bbsSn );
